+++ tests/integration/conversations/v1/service/test_user.py
@@ -47,7 +47,8 @@ class UserTestCase(IntegrationTestCase):
                 "identity": "admin",
                 "friendly_name": "name",
                 "attributes": "{ \\"duty\\": \\"tech\\" }",
-                "is_online": null,
+                "is_online": true,
+                "is_notifiable": null,
                 "date_created": "2019-12-16T22:18:37Z",
                 "date_updated": "2019-12-16T22:18:38Z",
                 "url": "https://conversations.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -86,7 +87,8 @@ class UserTestCase(IntegrationTestCase):
                 "identity": "admin",
                 "friendly_name": "new name",
                 "attributes": "{ \\"duty\\": \\"tech\\", \\"team\\": \\"internals\\" }",
-                "is_online": null,
+                "is_online": true,
+                "is_notifiable": null,
                 "date_created": "2019-12-16T22:18:37Z",
                 "date_updated": "2019-12-16T22:18:38Z",
                 "url": "https://conversations.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -148,7 +150,8 @@ class UserTestCase(IntegrationTestCase):
                 "identity": "admin",
                 "friendly_name": "name",
                 "attributes": "{ \\"duty\\": \\"tech\\" }",
-                "is_online": null,
+                "is_online": true,
+                "is_notifiable": null,
                 "date_created": "2019-12-16T22:18:37Z",
                 "date_updated": "2019-12-16T22:18:38Z",
                 "url": "https://conversations.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -196,7 +199,8 @@ class UserTestCase(IntegrationTestCase):
                         "identity": "admin",
                         "friendly_name": "name",
                         "attributes": "{ \\"duty\\": \\"tech\\" }",
-                        "is_online": null,
+                        "is_online": true,
+                        "is_notifiable": null,
                         "date_created": "2019-12-16T22:18:37Z",
                         "date_updated": "2019-12-16T22:18:38Z",
                         "url": "https://conversations.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -210,6 +214,7 @@ class UserTestCase(IntegrationTestCase):
                         "friendly_name": "John from customs",
                         "attributes": "{ \\"duty\\": \\"agent\\" }",
                         "is_online": false,
+                        "is_notifiable": null,
                         "date_created": "2020-03-24T20:38:21Z",
                         "date_updated": "2020-03-24T20:38:21Z",
                         "url": "https://conversations.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"