+++ tests/integration/conversations/v1/test_user.py
@@ -46,7 +46,8 @@ class UserTestCase(IntegrationTestCase):
                 "identity": "admin",
                 "friendly_name": "name",
                 "attributes": "{ \\"duty\\": \\"tech\\" }",
-                "is_online": null,
+                "is_online": true,
+                "is_notifiable": null,
                 "date_created": "2019-12-16T22:18:37Z",
                 "date_updated": "2019-12-16T22:18:38Z",
                 "url": "https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -83,7 +84,8 @@ class UserTestCase(IntegrationTestCase):
                 "identity": "admin",
                 "friendly_name": "new name",
                 "attributes": "{ \\"duty\\": \\"tech\\", \\"team\\": \\"internals\\" }",
-                "is_online": null,
+                "is_online": true,
+                "is_notifiable": null,
                 "date_created": "2019-12-16T22:18:37Z",
                 "date_updated": "2019-12-16T22:18:38Z",
                 "url": "https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -141,7 +143,8 @@ class UserTestCase(IntegrationTestCase):
                 "identity": "admin",
                 "friendly_name": "name",
                 "attributes": "{ \\"duty\\": \\"tech\\" }",
-                "is_online": null,
+                "is_online": true,
+                "is_notifiable": null,
                 "date_created": "2019-12-16T22:18:37Z",
                 "date_updated": "2019-12-16T22:18:38Z",
                 "url": "https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -187,7 +190,8 @@ class UserTestCase(IntegrationTestCase):
                         "identity": "admin",
                         "friendly_name": "name",
                         "attributes": "{ \\"duty\\": \\"tech\\" }",
-                        "is_online": null,
+                        "is_online": true,
+                        "is_notifiable": null,
                         "date_created": "2019-12-16T22:18:37Z",
                         "date_updated": "2019-12-16T22:18:38Z",
                         "url": "https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
@@ -201,6 +205,7 @@ class UserTestCase(IntegrationTestCase):
                         "friendly_name": "John from customs",
                         "attributes": "{ \\"duty\\": \\"agent\\" }",
                         "is_online": false,
+                        "is_notifiable": null,
                         "date_created": "2020-03-24T20:38:21Z",
                         "date_updated": "2020-03-24T20:38:21Z",
                         "url": "https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"